Allow userspace control of runtime disabling/enabling of driver probing
Various attacks are made possible due to the large attack surface of kernel
drivers and the easy availability of hotpluggable hardware that can be
programmed to mimic arbitrary devices. This allows attackers to find a
single vulnerable driver and then produce a device that can exploit it
by plugging into a hotpluggable bus (such as PCI or USB). This violates user
assumptions about unattended systems being secure as long as the screen is
locked.

The kernel already has support for deferring driver binding in order to
avoid problems over suspend/resume. By exposing this to userspace we can
disable probing when the screen is locked and simply reenable it on unlock.

This is not a complete solution - since this still permits device creation
and simply blocks driver binding, it won't stop userspace drivers from
attaching to devices and it won't protect against any kernel vulnerabilities
in the core bus code. However, it should be sufficient to block attacks like
Poisontap (https://samy.pl/poisontap/).

Signed-off-by: Matthew Garrett <mjg59@coreos.com>