IMA: Allow EVM validation on appraisal even without a symmetric key
A reasonable configuration is to use IMA to appraise a subset of files
(based on user, security label or other features supported by IMA) but
to also want to use EVM to validate not only the state of the IMA hash
but also additional metadata on the file. Right now this is only
possible if a symmetric key has been loaded, which may not be desirable
in all cases (eg, one where EVM digital signatures are shipped to end
systems rather than EVM HMACs being generated locally). Add an
additional "require_evm" keyword to the IMA policy language in order to
permit the local admin to indicate that they wish EVM validation to
occur even if no symmetric key has been loaded.

Signed-off-by: Matthew Garrett <mjg59@google.com>