amiga: Fix DMAC bounce buffer handling
The DMAC can only address 24 bits, so buffers allocated by the block core
may be unreachable. The existing logic attempts to fall back to allocating
a new block, but that may not be sufficient - m68k doesn't guarantee that
GFP_DMA will be below 24 bits, and there may not be *any* fast RAM below
24 bits in any case. However, as long as we're not a Draco, we're
guaranteed to have some chip RAM in that range. Add a fallback to that,
and keep the allocation around for reuse if possible rather than
generating additional churn.

Signed-off-by: Matthew Garrett <mjg59@google.com>