amiga: Bind DMAC driver to the Zorro device
The Zorro device ID used by the A2091 driver actually refers to the DMAC
rather than to the SCSI controller. Since CDTVs use the same DMAC and
Zorro ID but don't necessarily have a SCSI controller, move the binding
to the DMAC driver and have that call out to the driver code.

Signed-off-by: Matthew Garrett <mjg59@google.com>