Allow in-kernel use of tpm_request_locality
TPMs support multiple localities, roughly analogous to CPU protection
levels. Higher localities are restricted for use in early firmware, but
localities 0-2 should be accessible at runtime. Currently we only make
use of locality 0, but locality 2 is envisaged for use by OS kernels. We
can make use of this for various things, such as being able to reset PCRs
that are intended for use by the OS without userland being able to mimic
this behaviour.

Signed-off-by: Matthew Garrett <mjgarrett@aurora.tech>