pm: hibernate: tie the encryption key to specific PCR values
The key blob is not secret, and by default the TPM will happily unseal
it regardless of system state. We can protect against that by sealing
the secret with a PCR policy - if the current PCR state doesn't match,
the TPM will refuse to release the secret. For now we'll just seal it
to PCR 23 (to ensure that the current TPM state was determined by the
kernel) and PCR 5 (to prove that the kernel supports restriction of
PCR 23). We can also then verify the PCR values from the point where
the key was created, in order to ensure that an attacker hasn't simply
generated their own key and put it in the header.

In the long term we may want a more flexible policy around this,
such as including PCR 7.

Signed-off-by: Matthew Garrett <mgarrett@aurora.tech>