efi: Add support for logging of kernel security features to PCR 5
The presence or absence of certain kernel security features may be relevant
for security policy decisions, either inside or outside the kernel. Add
support for measuring these into the TPM during boot in order to allow
for attestation of security properties or for secrets to be sealed such that
kernels with a different security feature set will be unable to access them.

PCR 5 is chosen because it's necessary for there to be an event in-between
the measurement occurring and userland starting. If there isn't, userland
can simply perform an equivalent extension itself and spoof the state. We
could cap the PCR ourselves after extending all security measurements, but
this could be avoided by simply rolling back to an older kernel. We
therefore need a PCR that we can guarantee will be extended after we have
the ability to run in-kernel code, even on kernels that don't implement this
feature.

PCR 5 fits that bill - according to spec, ExitBootServices() must cause an
extension of PCR 5. Measuring into PCR 5 therefore guarantees that we can
distinguish between an extension that occurred before ExitBootServices()
(ie, an extension that must have occurred in the boot stub or some other
part of the trusted boot chain) and an extension that occurred in userland.
In the event that the UEFI stub is skipped, the bootloader must have called
ExitBootServices() on our behalf, and again we will be able to distinguish
the states.

Signed-off-by: Matthew Garrett <mgarrett@aurora.tech>